+++ client/src/components/product/ProductCard.css
@@ -1,0 +1,37 @@
+.card {
+  background-color: var(--tertiary-color);
+}
+
+.cards {
+  color: var(--primary-color);
+  padding-right: 12px;
+}
+
+.cards .cards-card {
+  background-color: var(--tertiary-color);
+  width: 24rem;
+  max-width: 100%;
+}
+
+.cards .cards-card .card-card-img {
+  height: 70vh;
+}
+
+.cards .cards-card .cards-card-footer {
+  font-size: var(--font-size-paragraph);
+  padding-top: 15px;
+  background-color: black;
+}
+
+.cards .cards-card .cards-card-price-icons {
+  font-size: var(--font-size-paragraph);
+  padding: 15px 0px 15px 0px;
+}
+
+.cards-card-heart {
+  font-size: 1.8rem;
+}
+
+.cards-card-cart {
+  font-size: 2rem;
+}
