+++ client/src/pages/ProductList/ProductList.css
@@ -1,0 +1,37 @@
+.cards-title {
+  background-color: var(--secondary-color);
+  font-size: var(--font-size-h1);
+  height: 100px;
+  padding: 10px 0px 0px 10px;
+  margin-bottom: 20px;
+}
+
+.cards-pagination-btn {
+  padding: 5px 12px;
+  border: 1px solid #ccc;
+  border-radius: 4px;
+  margin: 15px 4px 0 4px;
+  cursor: pointer;
+  color: var(--primary-color);
+  transition: all 0.2s ease;
+}
+
+.cards-pagination-btn:hover {
+  background-color: var(--primary-color);
+  color: var(--btn-cta-secondary-color);
+  border-color: var(--primary-color);
+}
+
+.cards-pagination-btn-active {
+  background-color: var(--secondary-color);
+  border-color: var(--primary-color);
+  color: #1e1e1e;
+}
+
+.rc-pagination-item,
+.rc-pagination-item-active,
+.rc-pagination-item:focus,
+.rc-pagination-item:hover {
+  border-color: transparent;
+  box-shadow: none;
+}
