+++ client/src/router.tsx
@@ -1,7 +1,7 @@
 import { createBrowserRouter } from "react-router";
 import Home from "./pages/Home.tsx";
 
-import ProductList from "./pages/ProductList.tsx";
+import ProductList from "./pages/ProductList/ProductList.tsx";
 
 const router = createBrowserRouter([
   {
