+++ client/src/types/product.ts
@@ -1,0 +1,8 @@
+export type ProductType = {
+  id: number;
+  name: string;
+  description: string;
+  price: number;
+  category_id: number;
+  images: string[];
+};
