+++ server/src/modules/product/productActions.ts
@@ -1,15 +1,16 @@
 import type { RequestHandler } from "express";
-
+import { StatusCodes } from "http-status-codes";
 import productRepository from "./productRepository";
 
 const browse: RequestHandler = async (req, res, next) => {
   try {
     const products = await productRepository.findAll();
-    if (products == null) {
-      res.sendStatus(404);
-    } else {
-      res.json(products);
+
+    if (products === null) {
+      res.status(StatusCodes.NOT_FOUND);
     }
+
+    res.json(products);
   } catch (err) {
     next(err);
   }
