+++ server/src/modules/product/productRepository.ts
@@ -11,8 +11,7 @@ class ProductRepository {
 
     for (const product of productRows) {
       const [imageRows] = await databaseClient.query<Rows>(
-        "SELECT path FROM image WHERE product_id = ?",
-        [product.id],
+        `SELECT path FROM image WHERE product_id = ${product.id}`,
       );
       product.images = imageRows.map((img) => img.path);
     }
